LibGfx: Respect order of image data in TGAImageDecoderPlugin
Previously, the order of data was incorrectly based on x_origin and
y_origin, which are meant for where on the screen the image should be
placed (TarGA is an image format meant for graphic cards), instead of
Image descriptor bits 4 and 5, which specify the direction.

This fixes a problem where tga files generated with magick would render
up-side-down.